requirements.txt
asgiref==3.11.0
Django==5.1.15
django-tenant-users==2.2.1
django-tenants==3.7.0
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.5
typing_extensions==4.15.0
